pyproject.toml
[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[project]
name = "ingestion-orchestrator"
version = "0.1.0"
description = "Databricks-native file ingestion orchestrator with Delta Lake state control and idempotent dispatch."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
  { name = "Gabri / Codex" }
]
dependencies = [
  "pyspark>=3.5.0,<4.0.0",
  "delta-spark>=3.2.0,<4.0.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "pytest-cov>=5.0.0",
  "ruff>=0.11.0",
  "mypy>=1.11.0"
]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
addopts = "-q"

[tool.ruff]
line-length = 110
target-version = "py310"
builtins = ["display", "dbutils"]

[tool.mypy]
python_version = "3.10"
disallow_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
ignore_missing_imports = true
